// src/types/inspector-types.ts
export enum InspectorTab {
  GENERAL = 'general',
  DESIGN = 'design',
  ADVANCED = 'advanced',
  ANIMATION = 'animation',
  RESPONSIVE = 'responsive'
}

export enum DesignProperty {
  BACKGROUND = 'background',
  BORDER = 'border',
  SPACING = 'spacing',
  TYPOGRAPHY = 'typography',
  LAYOUT = 'layout',
  EFFECTS = 'effects'
}

export enum FlexDirection {
  ROW = 'row',
  ROW_REVERSE = 'row-reverse',
  COLUMN = 'column',
  COLUMN_REVERSE = 'column-reverse'
}

export enum FlexWrap {
  NOWRAP = 'nowrap',
  WRAP = 'wrap',
  WRAP_REVERSE = 'wrap-reverse'
}

export enum JustifyContent {
  FLEX_START = 'flex-start',
  FLEX_END = 'flex-end',
  CENTER = 'center',
  SPACE_BETWEEN = 'space-between',
  SPACE_AROUND = 'space-around',
  SPACE_EVENLY = 'space-evenly'
}

export enum AlignItems {
  STRETCH = 'stretch',
  FLEX_START = 'flex-start',
  FLEX_END = 'flex-end',
  CENTER = 'center',
  BASELINE = 'baseline'
}

export enum AlignContent {
  STRETCH = 'stretch',
  FLEX_START = 'flex-start',
  FLEX_END = 'flex-end',
  CENTER = 'center',
  SPACE_BETWEEN = 'space-between',
  SPACE_AROUND = 'space-around'
}